backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
